refactor: moved table row to seperate component
