+++ src/lib/list/Info.svelte
@@ -1,6 +1,5 @@
 <script>
     import {Card, Accordion, AccordionItem} from 'flowbite-svelte'
-    import InfoView from '../InfoView.svelte';
     import ContractView from '../ContractView.svelte';
     import {Selected, SelectedExpanded} from "../selected";
 
@@ -14,13 +13,11 @@
 
 <Card size="xl" class="m-2">
     <Accordion multiple class="w-60">
-
         {#each $Selected as item, i}
           <AccordionItem bind:open={$SelectedExpanded[i]}>
-            <div slot="header">{item.name}</div>
-            <InfoView {item} />
-            <br>
-            <InfoView />
+            <div slot="header">{item.ID}</div>
+            <!-- TODO: how are we going to find the type of the item? should we change the backend -->
+            <code>fetch GET /api/ITEMTYPE/{item.ID}</code>
           </AccordionItem>
         {/each}
       </Accordion>
